// src/app/modules/transportation/models/vehicle-Insurance-filter-view-model.ts
/**
 * Transport
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AttachmentViewModel } from './attachment-view-model';


/**
 * VehicalInsuranceViewModel
 */
export interface VehicleInsuranceFilterViewModel {
    /**
     * Gets or Sets PolicyNumber
     */
    policyNumber?: string;
    /**
     * Gets or Sets PremiumAmount
     */
    premiumAmount?: any;
    /**
     * Gets or Sets CompanyName
     */
    companyName?: string;
    /**
     * Gets or Sets IssueDate
     */
    issueDateBegin?: any;

     /**
     * Gets or Sets IssueDate
     */
    issueDateEnd?: any;
    /**
     * Gets or Sets ExpiryDate
     */
    expiryDateBegin?: any;
      /**
     * Gets or Sets ExpiryDate
     */
    expiryDateEnd?: any;
   
    /**
     * gets or sets VehicleId
     */
    vehicleId?: string;
    /**
         * gets or sets SortBy
         */
    sortBy?: string;
    /**
         * gets or sets SortOrder
         */
    sortOrder: number;
    /**
         * gets or sets PageNumber
         */
    pageNumber: number;
    /**
     * gets or sets PageSize
     */
    pageSize?: number;
}