// src/app/modules/transportation/models/vehicle-Insurance-view-model.ts
/**
 * Transport
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AttachmentViewModel } from './attachment-view-model';


/**
 * VehicalInsuranceViewModel
 */
export interface VehicleInsuranceViewModel { 
    /**
     * Gets or Sets PolicyNumber
     */
    policyNumber?: string;
    /**
     * Gets or Sets PremiumAmount
     */
    premiumAmount?: number;
    /**
     * Gets or Sets CompanyName
     */
    companyName?: string;
    /**
     * Gets or Sets IssueDate
     */
    issueDate?: Date;
    /**
     * Gets or Sets ExpiryDate
     */
    expiryDate?: Date;
    /**
     * gets or sets Attachment
     */
    attachment?: AttachmentViewModel;
     /**
     * gets or sets Attachment
     */
    id: string;
}